fixed some vistual

+++ app/controllers/procedures_controller.rb
@@ -4,12 +4,14 @@ class ProceduresController < ApplicationController
   get '/members/:id/procedures' do
 
     @member=Member.find_by_id(params[:id])
+    authorize(@member)
     erb :"/procedures/index"
   end
 
   # GET: /procedures/new
   get "/members/:id/procedures/new" do
       @member=Member.find_by_id(params[:id])
+      authorize(@member)
       @diseases=Helpers.all_desease(session)
       erb :"/procedures/new"
   end
@@ -19,6 +21,7 @@ class ProceduresController < ApplicationController
     
     diseases=Helpers.all_desease(session)
     member=Member.find_by_id(params[:id])
+    authorize(member)
     procedure=Procedure.create(name:params[:name],desc:params[:desc],proce_doctor:params[:proce_doctor],
     hospital:params[:hospital],date_porc:params[:date_porc],age_at_proc:params[:age_at_proc])
 
@@ -44,6 +47,7 @@ class ProceduresController < ApplicationController
 
   get "/members/:id/procedures/:id1/edit" do
        @member=Member.find_by_id(params[:id])
+       authorize(@member)
        @procedure = Procedure.find_by_id(params[:id1])
        @diseases=Helpers.all_desease(session)
        erb :"/procedures/edit"
@@ -53,6 +57,7 @@ class ProceduresController < ApplicationController
 
     diseases=Helpers.all_desease(session)
     member=Member.find_by_id(params[:id])
+    authorize(member)
     procedure=Procedure.find_by_id(params[:id1])
     procedure.update(name:params[:name],desc:params[:desc],proce_doctor:params[:proce_doctor],
     hospital:params[:hospital],date_porc:params[:date_porc],age_at_proc:params[:age_at_proc])
@@ -72,6 +77,7 @@ class ProceduresController < ApplicationController
 
   delete "/members/:id/procedures/:id1" do
     member=Member.find_by_id(params[:id])
+    authorize(member)
     procedure=Procedure.find_by_id(params[:id1])
     procedure.destroy
     redirect "/members/#{member.id}/procedures"
